delete hotel
delete hotel
